requirements.txt
# Google ADK Dependencies
# Install with: pip install -r requirements.txt

google-adk
